pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "query-lapse"
version = "0.1.0"
description = "Record DB queries and replay them as deterministic test fixtures with PII scrubbing"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "query-lapse contributors"}]
keywords = ["database", "testing", "fixtures", "pii", "anonymization", "sql"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Database",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "click>=8.0",
]

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9"]
all = ["psycopg2-binary>=2.9"]

[project.scripts]
query-lapse = "query_lapse.cli:cli"

[project.urls]
Homepage = "https://github.com/query-lapse/query-lapse"
Issues = "https://github.com/query-lapse/query-lapse/issues"

[tool.setuptools.packages.find]
include = ["query_lapse*"]
